requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
gunicorn==20.1.0
python-dotenv==0.21.1
supabase==1.0.3
pydantic[email]<3.0.0
pydantic-settings==2.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
mangum==0.17.0
